+++ gemlib/a_graf_shrinkbox.c
@@ -2,9 +2,9 @@
 #include "gem.h"
 
 
-int
-graf_shrinkbox (int Fx, int Fy, int Fw, int Fh, int Sx, int Sy, int Sw,
-		int Sh)
+short
+graf_shrinkbox (short Fx, short Fy, short Fw, short Fh, short Sx, short Sy, short Sw,
+		short Sh)
 {
 	aes_intin[0] = Fx;
 	aes_intin[1] = Fy;
@@ -14,11 +14,14 @@ graf_shrinkbox (int Fx, int Fy, int Fw, int Fh, int Sx, int Sy, int Sw,
 	aes_intin[5] = Sy;
 	aes_intin[6] = Sw;
 	aes_intin[7] = Sh;
+	
 	aes_control[0] = 74;
 	aes_control[1] = 8;
 	aes_control[2] = 1;
 	aes_control[3] = 0;
 	aes_control[4] = 0;
+	
 	aes (&aes_params);
+	
 	return aes_intout[0];
 }
